pubsub: implement on top of generated client
The pubsub client uses the generated client, and therefore gRPC.

This is not a breaking change, but it may have performance implications
if you create many pubsub.Clients. The current version of gRPC does not
reuse connections, so each Client manages its own TCP connection. You
should create clients sparingly and call the Close method on them when
done (although closing on program termination is not necessary).

The biggest change was in the iterators. I made the smallest changes
that kept everything working. There will be more changes soon as
the generated code adopts the latest iterator design, and that
trickles into this client.

Reviewed-on: https://code-review.googlesource.com/7911

+++ pubsub/puller.go
@@ -45,7 +45,7 @@ type puller struct {
 // newPuller constructs a new puller.
 // batchSize is the maximum number of messages to fetch at once.
 // No more than batchSize messages will be outstanding at any time.
-func newPuller(s service, subName string, ctx context.Context, batchSize int64, keepAlive, abandon func(ackID string)) *puller {
+func newPuller(s service, subName string, ctx context.Context, batchSize int32, keepAlive, abandon func(ackID string)) *puller {
 	ctx, cancel := context.WithCancel(ctx)
 	return &puller{
 		cancel:    cancel,
